Vuforia is much more developed, got to figure out exactly how to work out the positioning and how to set it all up

+++ TeamCode/src/main/java/org/virus/agobot/MecanumVectorDriveTrain.java
@@ -404,7 +404,7 @@ public class MecanumVectorDriveTrain extends Drivetrain {
         opMode.telemetry.addData("new Position", newPosition);
         opMode.telemetry.addData("Change in position (rotated)", deltaPos);
 
-        robotCentricMvmt = new Vector2D((double) xController.getValue((float) -deltaPos.getComponent(1).doubleValue()), (double) yController.getValue((float) deltaPos.getComponent(0).doubleValue()));
+        robotCentricMvmt = new Vector2D((double) xController.getValue((float) -deltaPos.getComponent(1)), (double) yController.getValue((float) deltaPos.getComponent(0).doubleValue()));
         steerMag = headingController.getValue((float)angleDifference(newHeading, getHeading()));
         opMode.telemetry.addData("Robot Centric Movement", robotCentricMvmt);
